chore: move archive helpers to a separate package
feat: initialize the rollapp through 'rollapp' cmd
refactor: remove interactive cmd
refactor: retrieve config values from the generated roller.toml file
refactor: disable tests as the flow changes
refactor: make config init functions public
