+++ cmd/config/config.go
@@ -4,7 +4,6 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/dymensionxyz/roller/cmd/config/export"
-	configInit "github.com/dymensionxyz/roller/cmd/config/init"
 	"github.com/dymensionxyz/roller/cmd/config/set"
 	"github.com/dymensionxyz/roller/cmd/config/show"
 )
@@ -14,7 +13,7 @@ func Cmd() *cobra.Command {
 		Use:   "config",
 		Short: "Commands for setting up and managing rollapp configuration files.",
 	}
-	cmd.AddCommand(configInit.InitCmd())
+
 	cmd.AddCommand(show.Cmd())
 	cmd.AddCommand(set.Cmd())
 	cmd.AddCommand(export.Cmd())
